Optimization & Improvements of Quality
-LoopeD Synthwave music
-Optimize Database
-RemoveD LevelStats Script

+++ Assets/Scripts/LeveLStat.cs
@@ -30,7 +30,7 @@ public class LevelStat
     public string LevelID { get; set; }
 
     public int TotalAttempts { get; set; }
-    public float TotalTime { get; set; }
+    public double TotalTime { get; set; }
 
     public bool StarCoin1 { get; set; }
     public bool StarCoin2 { get; set; }
